// src/app/core/error-handling/error-types.ts
export enum ErrorType {
  Forbidden = 'Forbidden',
  ValidationError = 'ValidationError',
  NotFound = 'NotFound',
  Unauthorized = 'Unauthorized',
  InvalidCredentials = 'InvalidCredentials',
  AlreadyExists = 'AlreadyExists',
  InvalidToken = 'InvalidToken',
  InvalidOperation = 'InvalidOperation',
  ServerUnavailable = 'ServerUnavailable',
}